backend/requirements.txt
djangorestframework
mysqlclient
pip